requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openpyxl==3.1.2
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.2
python-dateutil==2.8.2
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0